+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 1. `uhppoted-set-super-passwords` node.
 
+### Updated
+1. Fixed input/output labels for renamed nodes.
+
 
 ## [1.1.5](https://github.com/uhppoted/node-red-contrib-uhppoted/releases/tag/v1.1.5) - 2023-08-30