+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Unreleased
+## [1.1.6](https://github.com/uhppoted/node-red-contrib-uhppoted/releases/tag/v1.1.6) - 2023-12-01
 
 ### Added
 1. `uhppoted-set-door-passcodes` node.